third_party/blink/web_tests/external/wpt/webrtc/RTCDataChannel-id-expected.txt

This is a testharness.js-based test.
PASS DTLS client uses odd data channel IDs
PASS DTLS server uses even data channel IDs
FAIL In-band negotiation with a specific ID should not be allowed assert_unreached: Channel established with same ID using negotiated=false Reached unreachable code
FAIL Odd/even role should not be violated when mixing with negotiated channels assert_equals: Channel id must be null before DTLS role has been determined expected (object) null but got (number) 0
Harness: the test ran to completion.
